+++ fcli-core/fcli-fod/src/main/java/com/fortify/cli/fod/access_control/helper/FoDUserHelper.java
@@ -141,7 +141,7 @@ public class FoDUserHelper {
         FoDAppDescriptor appDescriptor = FoDAppHelper.getAppDescriptor(unirest, appNameOrId, true);
         if (action.equals(FoDEnums.UserApplicationAccessAction.Add)) {
             FoDUserAppAccessRequest appAccessRequest = FoDUserAppAccessRequest.builder()
-                    .applicationId(Integer.valueOf(appDescriptor.getApplicationId())).build();
+                    .applicationId(Long.valueOf(appDescriptor.getApplicationId())).build();
             ObjectNode body = objectMapper.valueToTree(appAccessRequest);
             unirest.post(FoDUrls.USER_APPLICATION_ACCESS).routeParam("userId", String.valueOf(userDescriptor.getUserId()))
                     .body(body).asString().getBody();
